+++ scripts/create_test_data.py
@@ -25,7 +25,7 @@ from fidesops.ops.models.privacy_request import (
     PrivacyRequestStatus,
 )
 from fidesops.ops.models.storage import ResponseFormat, StorageConfig
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.schemas.storage.storage import FileNaming, StorageDetails, StorageType
 from fidesops.ops.util.data_category import DataCategory
 
@@ -195,7 +195,7 @@ def create_test_data(db: orm.Session) -> FidesUser:
             )
             pr.persist_identity(
                 db=db,
-                identity=PrivacyRequestIdentity(
+                identity=Identity(
                     email="test@example.com",
                     phone_number="+1 234 567 8910",
                 ),