+++ tests/ops/integration_tests/saas/request_override/test_mailchimp_override_task.py
@@ -4,7 +4,7 @@ import pytest
 
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.privacy_request import ExecutionLog, PrivacyRequest
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.task import graph_task
 from fidesops.ops.task.graph_task import get_cached_data_for_erasures
 from tests.ops.graph.graph_test_util import assert_rows_match, records_matching_fields
@@ -42,7 +42,7 @@ async def test_mailchimp_override_access_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_saas_access_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": mailchimp_identity_email})
+    identity = Identity(**{"email": mailchimp_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = mailchimp_override_connection_config.get_saas_config().fides_key
@@ -156,7 +156,7 @@ async def test_mailchimp_erasure_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_saas_erasure_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": mailchimp_identity_email})
+    identity = Identity(**{"email": mailchimp_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = mailchimp_override_connection_config.get_saas_config().fides_key