+++ tests/ops/integration_tests/saas/test_adobe_campaign_task.py
@@ -5,7 +5,7 @@ import pytest
 from fidesops.ops.core.config import config
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.privacy_request import PrivacyRequest
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.service.connectors import get_connector
 from fidesops.ops.task import graph_task
 from fidesops.ops.task.graph_task import get_cached_data_for_erasures
@@ -35,7 +35,7 @@ async def test_adobe_campaign_access_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_adobe_campaign_access_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": adobe_campaign_identity_email})
+    identity = Identity(**{"email": adobe_campaign_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = adobe_campaign_connection_config.get_saas_config().fides_key
@@ -179,7 +179,7 @@ async def test_adobe_campaign_saas_erasure_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_saas_access_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": erasure_email})
+    identity = Identity(**{"email": erasure_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = adobe_campaign_connection_config.get_saas_config().fides_key