+++ tests/ops/integration_tests/saas/test_hubspot_task.py
@@ -5,7 +5,7 @@ import pytest
 from fidesops.ops.core.config import config
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.privacy_request import PrivacyRequest
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.service.connectors import get_connector
 from fidesops.ops.task import graph_task
 from fidesops.ops.task.filter_results import filter_data_categories
@@ -39,7 +39,7 @@ async def test_saas_access_request_task(
     identity_attribute = "email"
     identity_value = hubspot_identity_email
     identity_kwargs = {identity_attribute: identity_value}
-    identity = PrivacyRequestIdentity(**identity_kwargs)
+    identity = Identity(**identity_kwargs)
     privacy_request.cache_identity(identity)
 
     dataset_name = connection_config_hubspot.get_saas_config().fides_key
@@ -147,7 +147,7 @@ async def test_saas_erasure_request_task(
     )
     identity_attribute = "email"
     identity_kwargs = {identity_attribute: (hubspot_erasure_identity_email)}
-    identity = PrivacyRequestIdentity(**identity_kwargs)
+    identity = Identity(**identity_kwargs)
     privacy_request.cache_identity(identity)
 
     dataset_name = connection_config_hubspot.get_saas_config().fides_key