+++ tests/ops/integration_tests/saas/test_salesforce_task.py
@@ -6,7 +6,7 @@ import requests
 from fidesops.ops.core.config import config
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.privacy_request import PrivacyRequest
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.service.connectors import get_connector
 from fidesops.ops.task import graph_task
 from fidesops.ops.task.graph_task import get_cached_data_for_erasures
@@ -36,7 +36,7 @@ async def test_salesforce_access_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_salesforce_access_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": salesforce_identity_email})
+    identity = Identity(**{"email": salesforce_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = salesforce_connection_config.get_saas_config().fides_key
@@ -395,7 +395,7 @@ async def test_salesforce_erasure_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_salesforce_erasure_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": salesforce_erasure_identity_email})
+    identity = Identity(**{"email": salesforce_erasure_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = salesforce_connection_config.get_saas_config().fides_key