+++ tests/ops/integration_tests/saas/test_sendgrid_task.py
@@ -5,7 +5,7 @@ import pytest
 from fidesops.ops.core.config import config
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.privacy_request import PrivacyRequest
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.task import graph_task
 from fidesops.ops.task.graph_task import get_cached_data_for_erasures
 from tests.ops.fixtures.saas.sendgrid_fixtures import contact_exists
@@ -27,7 +27,7 @@ async def test_sendgrid_access_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_saas_access_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": sendgrid_identity_email})
+    identity = Identity(**{"email": sendgrid_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = sendgrid_connection_config.get_saas_config().fides_key
@@ -85,7 +85,7 @@ async def test_sendgrid_erasure_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_saas_erasure_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": sendgrid_erasure_identity_email})
+    identity = Identity(**{"email": sendgrid_erasure_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = sendgrid_connection_config.get_saas_config().fides_key