+++ tests/ops/integration_tests/saas/test_stripe_task.py
@@ -7,7 +7,7 @@ import requests
 from fidesops.ops.core.config import config
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.privacy_request import PrivacyRequest
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.task import graph_task
 from fidesops.ops.task.filter_results import filter_data_categories
 from fidesops.ops.task.graph_task import get_cached_data_for_erasures
@@ -29,7 +29,7 @@ async def test_stripe_access_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_stripe_access_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": stripe_identity_email})
+    identity = Identity(**{"email": stripe_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = stripe_connection_config.get_saas_config().fides_key
@@ -650,7 +650,7 @@ async def test_stripe_erasure_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_stripe_erasure_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": stripe_erasure_identity_email})
+    identity = Identity(**{"email": stripe_erasure_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = stripe_connection_config.get_saas_config().fides_key