+++ tests/ops/integration_tests/saas/test_zendesk_task.py
@@ -7,7 +7,7 @@ import requests
 from fidesops.ops.core.config import config
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.privacy_request import PrivacyRequest
-from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
+from fidesops.ops.schemas.redis_cache import Identity
 from fidesops.ops.task import graph_task
 from fidesops.ops.task.graph_task import get_cached_data_for_erasures
 from tests.ops.graph.graph_test_util import assert_rows_match
@@ -28,7 +28,7 @@ async def test_zendesk_access_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_zendesk_access_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": zendesk_identity_email})
+    identity = Identity(**{"email": zendesk_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = zendesk_connection_config.get_saas_config().fides_key
@@ -194,7 +194,7 @@ async def test_zendesk_erasure_request_task(
     privacy_request = PrivacyRequest(
         id=f"test_zendesk_erasure_request_task_{random.randint(0, 1000)}"
     )
-    identity = PrivacyRequestIdentity(**{"email": zendesk_erasure_identity_email})
+    identity = Identity(**{"email": zendesk_erasure_identity_email})
     privacy_request.cache_identity(identity)
 
     dataset_name = zendesk_connection_config.get_saas_config().fides_key